Algorithms from pseudocode and exercises from book in description
Here are some algorithms written in Python either implementing pseudocode and solving exercises in the book Introduction to Algorithms by Cormen, Leiserson, Rivest, Stein, MIT press, 3rd Edition.

I sometimes use external library objects such as Numpy arrays (or even a Pandas data frame in one instance) which obviously are not very efficient in algorithms. I do this only for fun, but they are trivially substituted by lower level Python objects. It is trivial for instance, to use a nested list instead of a numpy.array to represent higher dimensional arrays. And if one wanted to be pedantic, one could even use the array from the ctypes library to get around the overhead of the extra space allocated to a dynamic array ADT. However, these details are completely useless for educational purposes, also because one would never apply algorithms written in python to industrial software, if looking for speed and reliability, but would rather use C/C++ or perhaps even Java.

One last comment, the comments in the code are scarce. This is just because I was too lazy to properly comment everything. However, it's straightforward to find relevant sections in the book; indeed,  the source code is supposed to be referring section explanations and exercises in the book.